+++ Tests/SentryTests/Integrations/Feedback/SentryFeedbackTests.swift
@@ -313,7 +313,8 @@ class SentryFeedbackTests: XCTestCase {
             timezone: try XCTUnwrap(TimeZone(identifier: "Europe/Vienna")),
             eventContextEnricher: TestEventContextEnricher(),
             crashWrapper: SentryDependencyContainer.sharedInstance().crashWrapper,
-            binaryImageCache: SentryDependencyContainer.sharedInstance().binaryImageCache
+            binaryImageCache: SentryDependencyContainer.sharedInstance().binaryImageCache,
+            dispatchQueueWrapper: TestSentryDispatchQueueWrapper()
         )
         let hub = TestHub(client: client, andScope: nil)
 
@@ -360,7 +361,8 @@ class SentryFeedbackTests: XCTestCase {
             timezone: try XCTUnwrap(TimeZone(identifier: "Europe/Vienna")),
             eventContextEnricher: TestEventContextEnricher(),
             crashWrapper: SentryDependencyContainer.sharedInstance().crashWrapper,
-            binaryImageCache: SentryDependencyContainer.sharedInstance().binaryImageCache
+            binaryImageCache: SentryDependencyContainer.sharedInstance().binaryImageCache,
+            dispatchQueueWrapper: TestSentryDispatchQueueWrapper()
         )
         let hub = TestHub(client: client, andScope: nil)
         SentrySDKInternal.setCurrentHub(hub)
@@ -407,7 +409,8 @@ class SentryFeedbackTests: XCTestCase {
             timezone: try XCTUnwrap(TimeZone(identifier: "Europe/Vienna")),
             eventContextEnricher: TestEventContextEnricher(),
             crashWrapper: SentryDependencyContainer.sharedInstance().crashWrapper,
-            binaryImageCache: SentryDependencyContainer.sharedInstance().binaryImageCache
+            binaryImageCache: SentryDependencyContainer.sharedInstance().binaryImageCache,
+            dispatchQueueWrapper: TestSentryDispatchQueueWrapper()
         )
         let hub = TestHub(client: client, andScope: nil)
         SentrySDKInternal.setCurrentHub(hub)